requirements.txt
discord.py
requests
os